+++ core/src/test/java/org/elasticsearch/index/similarity/SimilarityModuleTests.java
@@ -37,89 +37,5 @@ import java.util.Collections;
 
 public class SimilarityModuleTests extends ModuleTestCase {
 
-    public void testAddSimilarity() {
-        Settings indexSettings = Settings.settingsBuilder()
-                .put(IndexMetaData.SETTING_VERSION_CREATED, Version.CURRENT)
-                .put("index.similarity.my_similarity.type", "test_similarity")
-                .put("index.similarity.my_similarity.key", "there is a key")
-                .build();
-        SimilarityModule module = new SimilarityModule(IndexSettingsModule.newIndexSettings(new Index("foo"), indexSettings, Collections.EMPTY_LIST));
-        module.addSimilarity("test_similarity", (string, settings) -> new SimilarityProvider() {
-            @Override
-            public String name() {
-                return string;
-            }
 
-            @Override
-            public Similarity get() {
-                return new TestSimilarity(settings.get("key"));
-            }
-        });
-        assertInstanceBinding(module, SimilarityService.class, (inst) -> {
-            if (inst instanceof SimilarityService) {
-                assertNotNull(inst.getSimilarity("my_similarity"));
-                assertTrue(inst.getSimilarity("my_similarity").get() instanceof TestSimilarity);
-                assertEquals("my_similarity", inst.getSimilarity("my_similarity").name());
-                assertEquals("there is a key" , ((TestSimilarity)inst.getSimilarity("my_similarity").get()).key);
-                return true;
-            }
-            return false;
-        });
-    }
-
-    public void testSetupUnknownSimilarity() {
-        Settings indexSettings = Settings.settingsBuilder()
-                .put("index.similarity.my_similarity.type", "test_similarity")
-                .put(IndexMetaData.SETTING_VERSION_CREATED, Version.CURRENT)
-                .build();
-        SimilarityModule module = new SimilarityModule(IndexSettingsModule.newIndexSettings(new Index("foo"), indexSettings, Collections.EMPTY_LIST));
-        try {
-            assertInstanceBinding(module, SimilarityService.class, (inst) -> inst instanceof SimilarityService);
-        } catch (IllegalArgumentException ex) {
-            assertEquals("Unknown Similarity type [test_similarity] for [my_similarity]", ex.getMessage());
-        }
-    }
-
-
-    public void testSetupWithoutType() {
-        Settings indexSettings = Settings.settingsBuilder()
-                .put("index.similarity.my_similarity.foo", "bar")
-                .put(IndexMetaData.SETTING_VERSION_CREATED, Version.CURRENT)
-                .build();
-        SimilarityModule module = new SimilarityModule(IndexSettingsModule.newIndexSettings(new Index("foo"), indexSettings, Collections.EMPTY_LIST));
-        try {
-            assertInstanceBinding(module, SimilarityService.class, (inst) -> inst instanceof SimilarityService);
-        } catch (IllegalArgumentException ex) {
-            assertEquals("Similarity [my_similarity] must have an associated type", ex.getMessage());
-        }
-    }
-
-
-    private static class TestSimilarity extends Similarity {
-        private final Similarity delegate = new BM25Similarity();
-        private final String key;
-
-
-        public TestSimilarity(String key) {
-            if (key == null) {
-                throw new AssertionError("key is null");
-            }
-            this.key = key;
-        }
-
-        @Override
-        public long computeNorm(FieldInvertState state) {
-            return delegate.computeNorm(state);
-        }
-
-        @Override
-        public SimWeight computeWeight(CollectionStatistics collectionStats, TermStatistics... termStats) {
-            return delegate.computeWeight(collectionStats, termStats);
-        }
-
-        @Override
-        public SimScorer simScorer(SimWeight weight, LeafReaderContext context) throws IOException {
-            return delegate.simScorer(weight, context);
-        }
-    }
 }
